This is a further round of work on floating point. I have re-worked the CSL implementation of lessp and greaterp, and prepared the ground for other arithmetic operations. In the process the support for various floating point formats has been moved forward, but perhaps more importantly I believe I have fixes¥d some long-stanidng bugs involving comparisons between floating point and big integer values. In this version if you try to read in a long precision float such as 1.0L0 into Lisp you will get a load of silly trace output which is there because at present it shows some rounding errors during reading and some formatting mistakes while printing the 128-bit long float values. But I have put in enought updates that I want to cement those with a check-in and then get back to understanding the long-float mistakes and fixing them as well as reviewing and enhancing other arithmetic. There is also now a bariable gcknt!* in CSL that is intended to do what the same variable does for PSL.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3945 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/syscsl.h
@@ -397,7 +397,7 @@ extern int terminal_eof_seen;
 // been spent and how many garbage collections have been done.
 //
 extern void report_time(int32_t t, int32_t gct);
-extern void report_space(int gccount, double percent);
+extern void report_space(uint64_t gccount, double percent);
 //
 // pause_for_user() gets called right at the end to give a chance for the
 // system to delay before closing the main output window.
